REFACTOR: move all code out of `lib.rs`
